added expenses menu
